completed app

+++ controllers/courseController.js
@@ -5,7 +5,7 @@ const departments = ['Math', 'English', 'Music', 'Art', 'PE', 'World Languages',
 module.exports.viewAll = async function(req,res){
     const courses = await Course.findAll();
     res.render('course/view_all', {courses});
-}
+};
 
 //profile
 module.exports.viewProfile= async function(req,res){
@@ -20,21 +20,29 @@ module.exports.viewProfile= async function(req,res){
         }
     }
     res.render('course/profile', {course, availableStudents})
-}
+};
 
 //render add form
 module.exports.renderAddForm =  function(req, res){
+    if (!req.user.can('add course')){
+        res.redirect('/');
+        return
+    }
     const course =  {
         name: '',
         department: departments[0],
         instructor_name: '',
         description:''
-    }
+    };
     res.render('course/add', {course, departments});
-}
+};
 
 //add
 module.exports.addCourse = async function(req, res){
+    if (!req.user.can('add course')){
+        res.redirect('/');
+        return
+    }
     const course = await Course.create({
         name: req.body.name,
         department: req.body.department,
@@ -42,16 +50,24 @@ module.exports.addCourse = async function(req, res){
         description: req.body.description
     });
     res.redirect(`/courses/profile/${course.id}`);
-}
+};
 
 //render edit form
 module.exports.renderEditForm = async function(req, res){
+    if (!req.user.can('edit course')){
+        res.redirect('/');
+        return
+    }
     const course = await Course.findByPk(req.params.id);
     res.render('course/edit', {course, departments});
-}
+};
 
 //update
 module.exports.updateCourse = async function(req, res){
+    if (!req.user.can('edit course')){
+        res.redirect('/');
+        return
+    }
     const course = await Course.update({
         name: req.body.name,
         department: req.body.department,
@@ -63,29 +79,49 @@ module.exports.updateCourse = async function(req, res){
         }
         });
     res.redirect(`/courses/profile/${req.params.id}`);
-}
+};
 
 //delete
 module.exports.deleteCourse = async function(req, res){
+    if (!req.user.can('delete course')){
+        res.redirect('/');
+        return
+    }
     await Course.destroy({
         where: {
             id:req.params.id
         }
         });
     res.redirect('/courses');
-}
+};
 
 //Add student to a course
 module.exports.enrollStudent = async function(req, res){
+    const isAdmin = req.user.can('enroll student');
+    const profileBelongsToUser = req.user.can('enroll self') && req.user.matchesStudentId(req.body.student);
+
+    //if they aren't staff and they aren't the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
     await StudentCourses.create({
         student_id: req.body.student,
         course_id: req.params.courseId
     });
     res.redirect(`/courses/profile/${req.params.courseId}`)
-}
+};
 
 //remove a student from a course
 module.exports.removeStudent = async function(req, res){
+    const isAdmin = req.user.can('drop student');
+    const profileBelongsToUser = req.user.can('drop self') && req.user.matchesStudentId(req.params.studentId);
+
+    //if they aren't staff and they aren't the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
     await StudentCourses.destroy({
         where: {
             course_id: req.params.courseId,
@@ -93,7 +129,7 @@ module.exports.removeStudent = async function(req, res){
         }
     });
     res.redirect(`/courses/profile/${req.params.courseId}`);
-}
+};
 
 function courseHasStudent(course, student) {
     for (let i = 0; i < course.students.length; i++) {
@@ -102,4 +138,4 @@ function courseHasStudent(course, student) {
         }
     }
     return false
-}
+};
