+++ controllers/studentController.js
@@ -2,12 +2,24 @@ const {Student, Course, StudentCourses} = require('../models')
 
 //view all
 module.exports.viewAll = async function (req, res) {
+    if (!req.user.can('view students')){
+        res.redirect('/');
+        return
+    }
     const students = await Student.findAll();
     res.render('student/view_all', {students});
-}
+};
 
 //profile
 module.exports.viewProfile = async function (req, res) {
+    const isAdmin = req.user.can('view students');
+    const profileBelongsToUser = req.user.can('view self') && req.user.matchesStudentId(req.params.id);
+
+    //if they aren't staff and they aren't the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
     const student = await Student.findByPk(req.params.id, {
         include: 'courses'
     });
@@ -19,37 +31,33 @@ module.exports.viewProfile = async function (req, res) {
         }
     }
     res.render('student/profile', {student, availableCourses})
-}
-
-//render add
-module.exports.renderAddForm = function (req, res) {
-    const student = {
-        first_name: '',
-        last_name: '',
-        grade_level: 9,
-    }
-    res.render('student/add', {student});
-}
-
-//add
-module.exports.addStudent = async function (req, res) {
-    const student = await Student.create({
-        first_name: req.body.first_name,
-        last_name: req.body.last_name,
-        grade_level: req.body.grade_level
-    });
-    res.redirect(`/students/profile/${student.id}`);
-}
+};
 
 //render edit
 module.exports.renderEditForm = async function (req, res) {
+    const isAdmin = req.user.can('edit student');
+    const profileBelongsToUser = req.user.can('edit self') && req.user.matchesStudentId(req.params.id);
+
+    //if they aren't staff and they aren't the student whose profile they are trying to  view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
     const student = await Student.findByPk(req.params.id);
     console.log(student);
     res.render('student/edit', {student});
-}
+};
 
 //update
 module.exports.updateStudent = async function (req, res) {
+    const isAdmin = req.user.can('edit student');
+    const profileBelongsToUser = req.user.can('edit self') && req.user.matchesStudentId(req.params.id);
+
+    //if they aren't staff and they aren't the student whose profile they are trying to  view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
     const student = await Student.update({
         first_name: req.body.first_name,
         last_name: req.body.last_name,
@@ -61,31 +69,50 @@ module.exports.updateStudent = async function (req, res) {
         }
     });
     res.redirect(`/students/profile/${req.params.id}`);
-}
+};
 
 //delete
 module.exports.deleteStudent = async function (req, res) {
+    if (!user.can('delete student')) {
+        res.redirect('/');
+        return
+    }
     await Student.destroy({
         where: {
             id: req.params.id
         }
     });
     res.redirect('/students');
-}
+};
 
 //Add course to student
 module.exports.enrollStudent = async function (req, res) {
+    const isAdmin = req.user.can('enroll student');
+    const profileBelongsToUser = req.user.can('enroll self') && req.user.matchesStudentId(req.params.studentId);
 
+    //if they aren't staff and they aren't the student whose profile they are trying to  view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
         await StudentCourses.create({
             student_id: req.params.studentId,
             course_id: req.body.course
-        })
+        });
         res.redirect(`/students/profile/${req.params.studentId}`);
 
-}
+};
 
 //delete course from student
 module.exports.removeCourse = async function(req, res){
+    const isAdmin = req.user.can('drop student');
+    const profileBelongsToUser = req.user.can('drop self') && req.user.matchesStudentId(req.params.studentId);
+
+    //if they aren't staff and they aren't the student whose profile they are trying to  view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
     await StudentCourses.destroy({
         where: {
             student_id: req.params.studentId,
@@ -93,7 +120,7 @@ module.exports.removeCourse = async function(req, res){
         }
     });
     res.redirect(`/students/profile/${req.params.studentId}`)
-}
+};
 
 function studentHasCourse(student, course) {
     for (let i = 0; i < student.courses.length; i++) {
